+++ internal/describe/system_microservice.go
@@ -64,11 +64,8 @@ func (exe *systemMicroserviceExecutor) Execute() error {
 			Namespace: exe.namespace,
 			Name:      exe.name,
 		},
-		Spec: yamlMsvc,
-		Status: map[string]interface{}{
-			"status":     FormatMicroserviceStatus(status),
-			"execStatus": FormatMicroserviceExecStatus(execStatus),
-		},
+		Spec:   yamlMsvc,
+		Status: FormatMicroserviceDescribeStatus(status, execStatus),
 	}
 
 	if exe.filename == "" {